requirements_enhanced_rag.txt
# Enhanced Working RAG Backend Requirements
# Core FastAPI dependencies (REQUIRED)
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# Environment and utilities (REQUIRED)
python-dotenv>=1.0.0

# Development and testing (REQUIRED)
requests>=2.31.0

# Optional: Remove pathlib2 as it's not needed for Python 3.4+
# pathlib2>=2.3.7

# Optional RAG dependencies (install if you want full RAG functionality)
# Uncomment these lines and run: pip install -r requirements_enhanced_rag.txt

# LlamaIndex core
# llama-index>=0.9.0
# llama-index-core>=0.9.0

# LlamaIndex integrations
# llama-index-llms-openai>=0.1.0
# llama-index-embeddings-openai>=0.1.0
# llama-index-vector-stores-pinecone>=0.1.0

# Vector database
# pinecone-client>=3.0.0

# Other dependencies
# openai>=1.0.0
# numpy>=1.24.0
# pandas>=2.0.0
